# backend/tests/openapi.snapshot.json
{
 "components": {
  "schemas": {
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "title": "Detail",
      "type": "array"
     }
    },
    "title": "HTTPValidationError",
    "type": "object"
   },
...
   "JobInfo": {
    "description": "API-41 payload.",
    "properties": {
     "created_at": {
      "title": "Created At",
      "type": "string"
     },
     "done": {
      "default": 0,
      "title": "Done",
      "type": "integer"
     },
     "error": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Error"
     },
     "eta_s": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Eta S"
     },
     "finished_at": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Finished At"
     },
     "job_id": {
      "title": "Job Id",
      "type": "string"
     },
     "key": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Key"
     },
     "kind": {
      "enum": [
       "index",
       "hash",
       "thumbnail",
       "radiomics",
       "mesh",
       "task",
       "open-convert"
      ],
      "title": "Kind",
      "type": "string"
     },
     "project_id": {
      "title": "Project Id",
      "type": "string"
     },
     "ref": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Ref"
     },
     "started_at": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Started At"
     },
     "status": {
      "enum": [
       "queued",
       "waiting_for_runner",
       "running",
       "succeeded",
       "failed",
       "cancelled",
       "interrupted"
      ],
      "title": "Status",
      "type": "string"
     },
     "total": {
      "default": 0,
      "title": "Total",
      "type": "integer"
     }
    },
    "required": [
     "job_id",
     "kind",
     "project_id",
     "status",
     "created_at"
    ],
    "title": "JobInfo",
    "type": "object"
   },
...
   "ValidationError": {
    "properties": {
     "ctx": {
      "title": "Context",
      "type": "object"
     },
     "input": {
      "title": "Input"
     },
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "title": "Location",
      "type": "array"
     },
     "msg": {
      "title": "Message",
      "type": "string"
     },
     "type": {
      "title": "Error Type",
      "type": "string"
     }
    },
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError",
    "type": "object"
   },
...
  }
 },
 "info": {
  "title": "Radiology Workbench",
  "version": "3.0.0.dev0"
 },
 "openapi": "3.1.0",
 "paths": {
...
  "/api/v1/jobs/{job_id}": {
   "get": {
    "operationId": "get_job_api_v1_jobs__job_id__get",
    "parameters": [
     {
      "in": "path",
      "name": "job_id",
      "required": true,
      "schema": {
       "title": "Job Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/JobInfo"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Get Job",
    "tags": [
     "jobs"
    ]
   }
  },
...
  "/api/v1/projects/{pid}/items/{iid}/image": {
   "get": {
    "operationId": "get_image_api_v1_projects__pid__items__iid__image_get",
    "parameters": [
     {
      "in": "path",
      "name": "pid",
      "required": true,
      "schema": {
       "title": "Pid",
       "type": "string"
      }
     },
     {
      "in": "path",
      "name": "iid",
      "required": true,
      "schema": {
       "title": "Iid",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/octet-stream": {}
      },
      "description": "Original file bytes"
     },
     "206": {
      "description": "Partial content (HTTP Range)"
     },
     "304": {
      "description": "Not modified (If-None-Match)"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Image bytes (Range, ETag)",
    "tags": [
     "volumes"
    ]
   }
  },
  "/api/v1/projects/{pid}/items/{iid}/mask": {
   "get": {
    "operationId": "get_mask_api_v1_projects__pid__items__iid__mask_get",
    "parameters": [
     {
      "in": "path",
      "name": "pid",
      "required": true,
      "schema": {
       "title": "Pid",
       "type": "string"
      }
     },
     {
      "in": "path",
      "name": "iid",
      "required": true,
      "schema": {
       "title": "Iid",
       "type": "string"
      }
     },
     {
      "in": "query",
      "name": "seg",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Seg"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/octet-stream": {}
      },
      "description": "Original file bytes"
     },
     "206": {
      "description": "Partial content (HTTP Range)"
     },
     "304": {
      "description": "Not modified (If-None-Match)"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Mask bytes of one segmentation set (Range, ETag; default `default_seg`)",
    "tags": [
     "volumes"
    ]
   }
  },
  "/api/v1/projects/{pid}/items/{iid}/mesh/{label}": {
   "get": {
    "description": "`seg` = segmentation set (ADR-0015); default `default_seg`.",
    "operationId": "get_mesh_api_v1_projects__pid__items__iid__mesh__label__get",
    "parameters": [
     {
      "in": "path",
      "name": "pid",
      "required": true,
      "schema": {
       "title": "Pid",
       "type": "string"
      }
     },
     {
      "in": "path",
      "name": "iid",
      "required": true,
      "schema": {
       "title": "Iid",
       "type": "string"
      }
     },
     {
      "in": "path",
      "name": "label",
      "required": true,
      "schema": {
       "title": "Label",
       "type": "integer"
      }
     },
     {
      "in": "query",
      "name": "smooth",
      "required": false,
      "schema": {
       "default": 1,
       "title": "Smooth",
       "type": "integer"
      }
     },
     {
      "in": "query",
      "name": "seg",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Seg"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/octet-stream": {}
      },
      "description": "gzip MZ3 mesh (world mm)"
     },
     "202": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/JobInfo"
        }
       }
      },
      "description": "Mesh job queued or running (`Location` = job)"
     },
     "304": {
      "description": "Not modified (If-None-Match)"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Surface mesh of one mask label (202 + job if not cached)",
    "tags": [
     "volumes"
    ]
   }
  },
  "/api/v1/projects/{pid}/items/{iid}/thumbnail": {
   "get": {
    "operationId": "get_thumbnail_api_v1_projects__pid__items__iid__thumbnail_get",
    "parameters": [
     {
      "in": "path",
      "name": "pid",
      "required": true,
      "schema": {
       "title": "Pid",
       "type": "string"
      }
     },
     {
      "in": "path",
      "name": "iid",
      "required": true,
      "schema": {
       "title": "Iid",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "image/webp": {}
      },
      "description": "WebP thumbnail"
     },
     "304": {
      "description": "Not modified (If-None-Match)"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "WebP thumbnail (404 until generated)",
    "tags": [
     "volumes"
    ]
   }
  },
...
 }
}